### 2018-03-01 fefbac2
improving paper for JOSS
