+++ paper.bib
@@ -65,7 +65,7 @@ doi = {10.5194/gmd-2017-193}
     author = {Edzer Pebesma},
     year = {2017}.
     note = {R package version 0.6-1},
-    url = {https://github.com/r-spatial/sf/},
+    url = {https://github.com/r-spatial/sf/}
 }
 
 @Manual{ncdf4,
@@ -73,7 +73,7 @@ doi = {10.5194/gmd-2017-193}
     author = {David Pierce},
     year = {2017},
     note = {R package version 1.16},
-    url = {https://CRAN.R-project.org/package=ncdf4},
+    url = {https://CRAN.R-project.org/package=ncdf4}
 }
 
 @Manual{raster,
@@ -120,5 +120,5 @@ doi = {10.5194/gmd-2017-193}
     organization = {R Foundation for Statistical Computing},
     address = {Vienna, Austria},
     year = {2017},
-    url = {https://www.R-project.org/},
+    url = {https://www.R-project.org/}
   }
