### 2018-03-01 #601 @ibarraespinosa
Dear @whedon @arfon @leeper  I have edited the paper.bib file. It now should compile PDF. Thanks
