+++ paper.bib
@@ -46,7 +46,7 @@ doi = {10.5194/gmd-2017-193}
     author = {Daniel Schuch},
     year = {2017},
     note = {R package version 0.664.1},
-    url = {https://github.com/atmoschem/EmissV},
+    url = {https://github.com/atmoschem/EmissV}
 }
 
 @article{freitas2011prep,
