+++ paper.bib
@@ -63,7 +63,7 @@ doi = {10.5194/gmd-2017-193}
 @Manual{sf,
     title = {sf: Simple Features for R},
     author = {Edzer Pebesma},
-    year = {2017}.
+    year = {2017},
     note = {R package version 0.6-1},
     url = {https://github.com/r-spatial/sf/}
 }
