+++ paper.bib
@@ -34,7 +34,6 @@ article{Ibarraetal2017,
 author = {Ibarra-Espinosa, S. and Ynoue, R. and O'Sullivan, S. and Pebesma, E. and Andrade, M. D. F. and Osses, M.},
 title = {VEIN v0.2.2: an R package for bottom-up Vehicular Emissions Inventories},
 journal = {Geoscientific Model Development Discussions},
-volume = {2017},
 year = {2017},
 pages = {1--29},
 url = {https://www.geosci-model-dev-discuss.net/gmd-2017-193/},
