+++ paper.md
@@ -31,7 +31,7 @@ Emissions are the pollutant mass released into the atmosphere [@pulles2010art]. 
 [@landrigan2017lancet].
 
 
-An important tool for policy decision is air quality models. They have been used not only to study the impact of different emissions scenarios for policy making but also to understand the dynamics of air pollutants in various parts of the world [@Andradeetal2015]. The inputs for an air quality models are meteorology and emissions. Currently, there are tools for developing emissions inventories such as the VEIN [@Ibarraetal2017] and the EmissV models [@emissv].  However, the existing tools for inputting the emissions into the air quality models are not written with a high-level language, such as PREP-Chem written in Fortran and C [@freitas2011prep]. Therefore, we developed **eixport**, a tool for doing the mentioned task, using R [@R], a high-level programming. 
+An important tool for policy decision is air quality models. They have been used not only to study the impact of different emissions scenarios for policy making but also to understand the dynamics of air pollutants in various parts of the world [@Andradeetal2015]. The inputs for an air quality models are meteorology and emissions. Currently, there are tools for developing emissions inventories such as the VEIN [@vein] and the EmissV models [@emissv].  However, the existing tools for inputting the emissions into the air quality models are not written with a high-level language, such as PREP-Chem written in Fortran and C [@freitas2011prep]. Therefore, we developed **eixport**, a tool for doing the mentioned task, using R [@R], a high-level programming. 
 
 eixport imports functions form the R packages sf [@sf] which provides functions for spatial vector data, providing bindings to the GDAL, GEOS, and Proj.4 C++ libraries. Also, eixport import functions from the package ncdf4 [@ncdf4], which interface to Unidata netCDF Format Data Files, and from the raster package [@raster], which provides functions to gridded data.
 
