+++ paper.bib
@@ -30,11 +30,13 @@
     doi={10.3389/fenvs.2015.00009},
 	publisher={Frontiers}
 }
-article{Ibarraetal2017,
+
+@article{vein,
 author = {Ibarra-Espinosa, S. and Ynoue, R. and O'Sullivan, S. and Pebesma, E. and Andrade, M. D. F. and Osses, M.},
 title = {VEIN v0.2.2: an R package for bottom-up Vehicular Emissions Inventories},
 journal = {Geoscientific Model Development Discussions},
 year = {2017},
+volume = {1},
 pages = {1--29},
 url = {https://www.geosci-model-dev-discuss.net/gmd-2017-193/},
 doi = {10.5194/gmd-2017-193}
