### 2018-03-06 #601 @leeper
@jhollist Would you be interested in reviewing this for JOSS?
