### 2018-03-06 #601 @leeper
@jhollist I don't think that's critical in this case given what the package does. But please don't feel obligated to say yes!
