### 2018-03-06 #607 @whedon
[REVIEW]: Eixport, an R Package to export emissions to atmospheric models
**Submitting author:** @ibarraespinosa (<a href="http://orcid.org/0000-0002-3162-1905">Sergio Ibarra</a>)
**Repository:** <a href="https://github.com/atmoschem/eixport" target ="_blank">https://github.com/atmoschem/eixport</a>
**Version:** v0.3.0
