### 2018-03-26 #607 @ibarraespinosa
Hi @leeper, thank you for the review. Could you please help me with some doubts regarding the unchecked boxes?:

_Authorship: Has the submitting author (@ibarraespinosa) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?_

I did not understand. Should I include/exclude authors?

_Functionality: Have the functional claims of the software been confirmed?_

The example on the paper and on the manual of eixport shows how to create inputs for WRFchem. In fact, I'm finishing a paper about air pollution modelling using WRF/Chem (Ibarra _et al_ 2018) whose inputs were created with eixport, as show Fig. 1 shows:

![fig1](https://user-images.githubusercontent.com/27447280/37936076-88cfd8a2-312a-11e8-8acc-1f6afc7d56c4.png)

The following figure (Fig 2) shows the results of the WRFChem air pollution simulation for the first 24 hours of CO (ppm). The results are set for a grid of 1 km of grid spacing.

![cowrf](https://user-images.githubusercontent.com/27447280/37937189-ee4d3bb2-312e-11e8-9ea9-93a16079983f.gif)

Ibarra-Espinosa S., Ynoue R., Giannotti M., Andrade MF., Shuch D., Vara Vela A., Gavidia-Calderon M. and Freitas E. (2018). Can we explain Ozone formation in a tropical mega-city? A multiple evaluation with dynamic emissions inventories from VEIN and air quality modeling with WRF-Chem. Paper in preparation for Science of Total Environment.

_Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?_

All the functions were tested before CRAN submissions. The tests are made in the examples of each R functions. As some of the examples took more than 5 seconds, I had to add \dontrun in r documentation, however, all checks passed and eixport [it is on CRAN now](https://CRAN.R-project.org/package=eixport). Moreover, eixport includes tests with [codecov](https://codecov.io/github/atmoschem/eixport?branch=master), however, as the eixport outputs are mostly heavy file files created at the disk, it is difficult to add more tests. For instance, the size of the file shown on Fig 1 (of this comment) has a size of 1.9 Gb. However, if you prefer @leeper,  i can send you the data for creating this file using eixport, with a script which follows the example shown on paper.

_Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support_

We have adhered to the code of conduct of JOSS, and we added a contributing section at README.Rmd file in [this](https://github.com/atmoschem/eixport/commit/6e272a6678e53ea87c81f3723ebf550d0add8258) commit.

_References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?_

After submissions to JOSS, we included a Zenodo DOI for eixport: [10.5281/zenodo.1186672](https://zenodo.org/record/1186672).

Many thanks,
Regards, Sergio
