+++ R/emis_opt.R
@@ -2,7 +2,7 @@
 #'
 #' @description  Emission package definitions from WRF 3.9.1.1
 #'
-#' @format A list of emision variables names, same number than emis_opt in namelist.
+#' @format A list of emision variables names, same number as emis_opt in namelist.
 #'
 #' @examples \dontrun{
 #' # Do not run
@@ -16,7 +16,7 @@
 #' }
 #'
 #' @note
-#' look to the number of aerosol of the emis_opt in wrf domumentation / code.
+#' look to the number of aerosol of the emis_opt in WRF domumentation / code.
 #'
 #' @seealso \code{\link{to_wrf}}
 #'
