### 2018-03-28 atmoschem/eixport#22 @jhollist
Suggested edits from Hollister review for JOSS
- [x] R Dependency to 3.3.0
- [x] Check on Authorship - Different author list on paper vs. software.
- [x] Fix example on `emis_opt`
- [x] Check example on `wrf_plot` to make sure the warning is not a problem
- [x] Add a separate section to README for installation instructions.  A good example is https://github.com/ropensci/lawn
- [x] Add tests.  Should have coverage for at least the main functions.
- [x] Add details on how to contribute in a CONTRIBUTING.md
- [x] Provide better justification for the package. In particular, provide more detail on why an R package is needed on top of the already available lower level implementation in Fortran and C.
- [x] Add DOIs to references.
- [x] Review [PR #21](https://github.com/atmoschem/eixport/pull/21) for suggested edits
- [ ] As a future enhancement, add a vignette showing typical use case
