### 2018-03-28 #607 @jhollist
# Hollister review for JOSS - Eixport R package

## Issue with checklist of edits

- [Issue #22](https://github.com/atmoschem/eixport/issues/22)

## Review notes

### General Comments

The R package, eixport, provides tools for importing and exporting files required for several emissions models.  It makes conversions as needed to allow greater integration between the variety of models.  Existing tools exist in lower level languages (e.g. Fortran or C) but this is the first implementation in R.

When issues in this review are addressed this will be a very nice contribution to JOSS.  As the authors address my (mostly minor) concerns they should pay close attention to the documentation.  There are a few places where it is thin and could be made more complete.  I have made some suggestions in a [Pull Request](https://github.com/atmoschem/eixport/pull/21) but in addition, the README should also be edited carefully.  Also, I have listed the most pressing issues in [Issue #22](https://github.com/atmoschem/eixport/issues/22) on the source repository

### Misc
- Depends set at >= 2.10 but imports are >= 3.3.0 and 3.0.0. That being said I don't think this issue has been solved in the larger community.  See [this discussion](https://community.rstudio.com/t/determining-which-version-of-r-to-depend-on/4396) for some details.  Having your dependency set to 2.10 implies that it works on that version which I am not sure is the case.  Were this my package I'd up to 3.3.0 which is not overly restrictive for most users.  Will leave this decision to the authors though.
- Vigentte that shows the typical use case and workflow might be a nice addition.  Not required for acceptance, but would be an enhacment for the future.  Captured in [Issue #22](https://github.com/atmoschem/eixport/issues/22).

### General Checks
- License is MIT, but full plain text file includes only copyright year/holder.  This is standard for R packages so I checked off on it.  Not sure if JOSS has additional requirements.
- Two authors on package, three on paper.  What were contributions from Edmilson Dias de Freitas?  It does appear that this work was inspired by the foundational work by Freitas.  Not sure what JOSS requires for authorship when differences between paper and software...

### Functionality

- example in `emis_opt` threw an error

    - Error in ncdf4::ncvar_def(name = variaveis[i], units = "", dim = list(west_east,  :
    - Passed a var name that is NOT a string of characters!

- example in `wrf_plot` gives a warning and plot appears to have no data

    - Warning message:
    - In wrf_plot(files[1], "E_CO") : Max value = Min Value!

### Documentation

- Installation instructions included on README, but probably better to have it in a separate section.  See https://github.com/ropensci/lawn for an example.
- Would like to see test beefed up a bit.  Currently has ~5% coverage an only covers the `to_rline` function.  Also, given that many of the examples are wrapped in \dontrun{} that means the standard testing done by CRAN won't find any potential problems.  Thus, a more complete set of tests are important.
- Code of conduct and  bug reports included in the DESCRIPTION.  I would like to see a CONTRIBUTING added that lays out guidelines for potential contributions.  A good example is https://github.com/ropensci/tidyhydat/blob/master/CONTRIBUTING.md  Just need something to direct potential contributors.

### Software paper

- I do think this is a valuable contribution and solves a problem, but as written the justification needs to be strengthened.  Currently the authors refer to the fact that existing tools are only available in lower level languages such as Fortran or C.  Need a bit more justification included in the paper as to why an implementation in R is desirable.  I agree that it is, just need to see it described in the paper.
- Several references are missing DOIs. For instance Freitas et al 2005 in Atmospheric Environment has a DOI of: 10.1016/j.atmosenv.2005.07.017  Include a DOI for all references that have them.

### Detailed edits

- I have suggested a number of edits in PR []() on the `eixport` repository.
- Line 6 in Lights.R needs more detail (or less, maybe just "A matrix" would work)
- Line 5 in to_brams_spm.R not sure what is meant by a function included for hourly profiles.
- Line 8 and 10 in to_brams_spm.R, What is meant by "polygon grid class 'sf'" Is this an `sf` object with MULTIPOLYGON geometry?
- Line 19 in to_brams_spm.R list it as SPM BRAMS, it is listed as BRAMS-SPM.  Be consistent how you refer to the models.
- Line 1 in wrf_profile.R is missing something.
