### 2018-03-28 #607 @leeper
@jhollist Great review. Thanks for your very careful examination of the package!!

@ibarraespinosa Please take a look through the review and try to respond to all points raised here and in https://github.com/atmoschem/eixport/issues/22.
