### 2018-03-28 #607 @ibarraespinosa
Thanks, @jhollist <https://github.com/jhollist>, and @leeper
<https://github.com/leeper>. I will read carefully and fix each point. Best
regards :)

Em qua, 28 de mar de 2018 às 15:14, Thomas J. Leeper <
notifications@github.com> escreveu:

> @jhollist <https://github.com/jhollist> Great review. Thanks for your
> very careful examination of the package!!
>
> @ibarraespinosa <https://github.com/ibarraespinosa> Please take a look
> through the review and try to respond to all points raised here and in
> atmoschem/eixport#22 <https://github.com/atmoschem/eixport/issues/22>.
>
