### 2018-03-29 8c6dde4
Functionality fix for JOSS
