+++ R/emis_opt.R
@@ -6,19 +6,15 @@
 #'
 #' @examples \dontrun{
 #' # Do not run
-#' dir.create("EMISS")
 #' data(emis_opt)
 #' names(emis_opt)
-#' wrf_create(wrfinput_dir = system.file("extdata", package = "eixport"),
-#'            wrfchemi_dir = "EMISS",
-#'            variaveis    = emis_opt["eradmsorg"],
-#'            n_aero       = 17)
+#' emis_opt["eradm"]
 #' }
 #'
 #' @note
 #' look to the number of aerosol of the emis_opt in wrf domumentation / code.
 #'
-#' @seealso \code{\link{to_wrf}}
+#' @seealso \code{\link{wrf_create}}
 #'
 #' @author Daniel Schuch
 #'
