### 2018-03-29 f29a181
Functionality fix for JOSS
