+++ paper.bib
@@ -58,6 +58,7 @@ doi = {10.5194/gmd-2017-193}
   number={2},
   pages={419},
   year={2011},
+  doi={10.5194/gmd-4-419-2011},
   publisher={Copernicus GmbH}
 }
 
@@ -91,6 +92,7 @@ doi = {10.5194/gmd-2017-193}
   journal={Atmospheric Environment},
   volume={39},
   pages={6352--6361},
+  doi={https://doi.org/10.1016/j.atmosenv.2005.07.017},
   year={2005}
 }
 
@@ -101,6 +103,7 @@ doi = {10.5194/gmd-2017-193}
   volume={77},
   pages={748--756},
   year={2013},
+  doi={https://doi.org/10.1016/j.atmosenv.2013.05.074},
   publisher={Elsevier}
 }
 
@@ -112,6 +115,7 @@ doi = {10.5194/gmd-2017-193}
   number={37},
   pages={6957--6975},
   year={2005},
+  doi={https://doi.org/10.1016/j.atmosenv.2005.04.027},
   publisher={Elsevier}
 }
 
