+++ man/to_brams_spm.Rd
@@ -2,7 +2,7 @@
 % Please edit documentation in R/to_brams_spm.R
 \name{to_brams_spm}
 \alias{to_brams_spm}
-\title{Inputs for BRAMS SPM}
+\title{Inputs for BRAMS-SPM}
 \usage{
 to_brams_spm(sdf, epsg = 4326)
 }
@@ -20,13 +20,13 @@ data-frame of daily gridded emissions, lat, long and a message with
 function.
 }
 \description{
-Create inputs for SPM brams. The inputs consists in data-frame
-or a lists of data-frames with daily emissions (mol/day), lat, long. Also,
+Create inputs for BRAMS-SPM. The inputs consist of a data-frame
+or a list of data-frames with daily emissions (mol/day), lat, long. Also,
 including a functions describing the hourly profile.
 }
 \note{
-When the input os class 'Spatial', they are converted to 'sf'. If the
-input is a data-frame, the output is a data-frame. If he input is a list,
+When the input is class 'Spatial', they are converted to 'sf'. If the
+input is a data-frame, the output is a data-frame. If the input is a list,
 the output is a list.
 }
 \examples{
