+++ DESCRIPTION
@@ -27,4 +27,4 @@ Imports: sf, ncdf4, raster, sp
 Encoding: UTF-8
 LazyData: no
 RoxygenNote: 6.0.1
-Suggests: testthat,covr
+Suggests: testthat, covr, cptcity
