### 2018-04-23 85de21a
add test wrf_create
