+++ README.Rmd
@@ -14,12 +14,9 @@ knitr::include_graphics("https://i.imgur.com/BcZ2tfW.png")
 
 Emissions are mass that affects atmosphere in complex ways, not only physical, but also, in the health of humans, ecosystems, economically, etc.
 
-There are several models whose inputs are emissions, such as [R-Line](https://www.cmascenter.org/r-line/) or [WRF-Chem](https://ruc.noaa.gov/wrf/wrf-chem/). Now experimental addition of inputs for running smoke and cmaq.
-
+There are several models whose inputs are emissions, such as [R-Line](https://www.cmascenter.org/r-line/) or [WRF-Chem](https://ruc.noaa.gov/wrf/wrf-chem/). 
 This R-Package provide functions to read emissions from [VEIN](https://github.com/ibarraespinosa/vein) and from other models 
 in different formats and export the emissions into the appropiate format suitable to other models.
-
-The first case of this package will exporting VEIN emissions into the model R-Line.
 
 
 ## Install
