+++ tests/testthat/test-to_as4wrf.R
@@ -7,8 +7,8 @@ test_that("to_as4wrf works", {
   expect_equal(df,
                data.frame(
                  id = 1,
-                 long =  -46.80416,
-                 lat = -23.61756,
+                 long =  -46.8066,
+                 lat = -23.62,
                  pollutant = 8391.2824346044,
                  time_lt = as.POSIXct(x = "2018-04-29", tz = "America/Sao_Paulo"),
                  time_utc = as.POSIXct(x = "2018-04-29 03:00:00", tz = "Etc/UTC"),
