+++ NAMESPACE
@@ -38,6 +38,4 @@ importFrom(sf,st_sf)
 importFrom(sf,st_sfc)
 importFrom(sf,st_transform)
 importFrom(sp,CRS)
-importFrom(sp,coordinates)
-importFrom(sp,spTransform)
 importFrom(utils,menu)
