+++ man/wrf_plot.Rd
@@ -5,7 +5,7 @@
 \title{Simple plot from wrf emission file}
 \usage{
 wrf_plot(file = file.choose(), name = NA, time = 1, nivel = 1,
-  barra = T, lbarra = 0.2, col = gray.colors(13), verbose = T, ...)
+  barra = T, lbarra = 0.2, col = cptcity::cpt(n = 13), verbose = T, ...)
 }
 \arguments{
 \item{file}{emission file name}
@@ -52,8 +52,6 @@ data(Lights)
 to_wrf(Lights, files[1], total = 1521983, names = "E_CO")
 
 wrf_plot(files[1], "E_CO")
-#library(cptcity) for nice colour scales
-#wrf_plot(files[1], "E_CO", col = cpt())
 }
 }
 \seealso{
