+++ NAMESPACE
@@ -11,6 +11,7 @@ export(wrf_plot)
 export(wrf_profile)
 export(wrf_put)
 import(ncdf4)
+importFrom(cptcity,cpt)
 importFrom(grDevices,cm.colors)
 importFrom(grDevices,gray.colors)
 importFrom(graphics,.filled.contour)
