+++ .travis.yml
@@ -27,25 +27,8 @@ before_install:
   - sudo apt-get install --yes libudunits2-dev
   # sf dependencies:
   - sudo apt-get install --yes libproj-dev libgeos-dev libgdal-dev libnetcdf-dev  netcdf-bin gdal-bin
-  # postgis source compile dependencies:
-  - sudo apt-get --yes install libjson-c-dev postgresql-server-dev-9.6
-  # install postgis from source:
-  - wget http://download.osgeo.org/postgis/source/postgis-2.3.2.tar.gz
-  - (mv postgis* /tmp; cd /tmp; tar xzf postgis-2.3.2.tar.gz)
-  - (cd /tmp/postgis-2.3.2 ; ./configure; make; sudo make install)
-  # create postgis databases:
-  - sudo service postgresql restart
-  - createdb postgis
-  - psql -d postgis -c "CREATE EXTENSION postgis;"
-  - psql -d postgis -c "GRANT CREATE ON DATABASE postgis TO travis"
-  - createdb empty
-  - psql -d empty -c "CREATE EXTENSION postgis;"
 
 warnings_are_errors: true
 
 after_success:
-  - dropdb postgis
-  - createdb postgis
-  - psql -d postgis -c "CREATE EXTENSION postgis;"
-  - psql -d postgis -c "GRANT CREATE ON DATABASE postgis TO travis"
   - Rscript -e 'covr::codecov()'
