### 2018-04-23 atmoschem/eixport#22 @ibarraespinosa
Dear @jhollist, with the our last [commit ](https://github.com/atmoschem/eixport/commit/badd00bf16af791105efcd4b4dad720cfe0493c3) we added coverage of 89%. It was not easy because but I resolved to add a file in inst/extdata for comparisons.

Please, let us know what do you think.

Best, Sergio
