+++ paper.md
@@ -37,22 +37,23 @@ eixport imports functions form the R packages sf [@sf] which provides functions 
 
 ## Functions and data
 
-eixport count with the folllwing functions:
+eixport count with the following functions:
 
 | Function     | Description                                                   |
 |--------------|---------------------------------------------------------------|
 | emisco       | Dataset of Emissions from VEIN demo                           |
 | emis_opt     | List of WRF emission species                                  |
 | rawprofile   | A matrix to temporally distribute emissions                   |
-| to_brams_spm | Produce inputs for BRAMS SPM [@freitas2005simple]             |
-| to_rline     | Produce inputs for R-Line [@snyder2013rline]                  |
-| to_wrf       | Distribution of emissions for WRF-Chem [@Grelletal2005]       |
 | wrf_create   | Create WRF-Chem inputs based on a WRFinput file               |
 | wrf_get      | Reads variables from WRF-Chem inputs                          |
 | wrf_grid     | Creates spatial feature (sf) polygon grid from WRFinput file  |
 | wrf_plot     | Simple plot from wrf emission file                            |
 | wrf_profile  | returns a traffic intensity profile (based on wrf file Times) |
 | wrf_put      | Function to write variables in WRF-Chem inputs                |
+| to_brams_spm | Produce inputs for BRAMS SPM [@freitas2005simple]             |
+| to_rline     | Produce inputs for R-Line [@snyder2013rline]                  |
+| to_wrf       | Distribution of emissions for WRF-Chem [@Grelletal2005]       |
+| to_as4wrf    | Produce data-frame to be used with NCL AS4WRF [@vara2016]     |
 
 ## Examples
 
@@ -61,30 +62,20 @@ create a wrfchem input in that file. The package already counts with wrfinput fi
 
 ```
 library(eixport)
-dir.create("EMISS")
+dir.create(file.path(tempdir(), "EMISS"))
 wrf_create(wrfinput_dir = system.file("extdata", package = "eixport"),
-          wrfchemi_dir = "EMISS",
-          frames_per_auxinput5 = 24)
+          wrfchemi_dir = file.path(tempdir(), "EMISS"))
 
 # get the name of created file
-files <- list.files(path = "EMISS",pattern = "wrfchemi",full.names = T)
+files <- list.files(path = file.path(tempdir(), "EMISS"),
+                   pattern = "wrfchemi",
+                   full.names = TRUE)
 
+# load end write some data in this emission file
 data(Lights)
+to_wrf(Lights, files[1], total = 1521983, names = "E_CO")
 
-perfil <- c(0.010760058, 0.005280596, 0.002883553, 0.002666932,
-           0.005781312, 0.018412838, 0.051900411, 0.077834636,
-           0.067919758, 0.060831614, 0.055852868, 0.052468599,
-           0.050938043, 0.051921718, 0.052756244, 0.052820165,
-           0.058388406, 0.072855890, 0.075267137, 0.063246412,
-           0.042713523, 0.029108975, 0.022091855, 0.015298458)
-
-plot(perfil,ty = "l",col= "purple",xlab = "Hour",main = "Time profile",
-    ylab = "Weight",axes = F,xlim = c(0,24))
-axis(2)
-axis(1,at = c(0,6,12,18,24),labels = c("00:00","06:00","12:00","18:00","00:00"))
-
-to_wrf(Lights,files[1],total = 1521983,profile = perfil,names = "E_CO")
-wrf_plot(files[1])
+wrf_plot(files[1], "E_CO")
 # [1] "EMISS/wrfchemi_d01_2011-08-01_00:00:00"
 # [1] "E_CO"
[…]
