+++ paper.bib
@@ -127,3 +127,21 @@ doi = {10.5194/gmd-2017-193}
     year = {2017},
     url = {https://www.R-project.org/}
   }
+@article{vara2016,
+  title={Impact of vehicular emissions on the formation of fine particles in the Sao Paulo Metropolitan Area: a numerical study with the WRF-Chem model},
+  author={Vara-Vela, A and Andrade, Maria F and Kumar, Prashant and Ynoue, Rita Y and Mu{\~n}oz, AG},
+  journal={Atmospheric Chemistry and Physics},
+  volume={16},
+  number={2},
+  pages={777--797},
+  year={2016},
+  publisher={Copernicus GmbH}
+}
+
+@manual{covr,
+    title = {covr: Test Coverage for Packages},
+    author = {Jim Hester},
+    year = {2017},
+    note = {R package version 3.0.1},
+    url = {https://CRAN.R-project.org/package=covr},
+  }
