+++ paper.bib
@@ -145,3 +145,10 @@ doi = {10.5194/gmd-2017-193}
     note = {R package version 3.0.1},
     url = {https://CRAN.R-project.org/package=covr},
   }
+@manual{cpt,
+    title = {cptcity: incorporating the cpt-city archive into R},
+    author = {Sergio Ibarra-Espinosa},
+    year = {2017},
+    note = {Colour gradients from http://soliton.vm.bytemark.co.uk/pub/cpt-city/},
+    url = {https://CRAN.R-project.org/package=vein},
+  }
