### 2018-04-24 #607 @jhollist
@ibarraespinosa Nice job on the tests!  89% is pretty good coverage and better than I have for most of my pacakges!

@leeper I give 👍 👍 and it looks good to go.
