### 2018-04-24 atmoschem/eixport#22 @jhollist
See comments on openjournals/joss-reviews#607, but the gist is 👍
