### 2018-04-24 #607 @ibarraespinosa
Thank you @jhollist. Suggesting codecov of code was very clever. It was difficult but it allowed me to improve the code.
