### 2018-04-25 #607 @leeper
Thank you @jhollist!!

@ibarraespinosa We're nearly there. Two final things:

 - Please add DOIs for all references that you can in [your bibliography](https://github.com/atmoschem/eixport/blob/master/paper.bib). While R packages do not have DOIs, most of the journal articles you are citing should have them.
 - Please update your Zenodo archive with the latest version of the package and let me know its DOI.
