+++ paper.bib
@@ -2,6 +2,7 @@
   title={The art of emission inventorying},
   author={Pulles, Tim and Heslinga, Dick},
   journal={TNO, Utrecht},
+  doi={10.13140/RG.2.1.2082.8007},
   year={2010}
 }
 
@@ -9,6 +10,7 @@
   title={Atmospheric chemistry and physics: from air pollution to climate change},
   author={Seinfeld, John H and Pandis, Spyros N},
   year={2016},
+  doi={10.1063/1.882420},
   publisher={John Wiley \& Sons}
 }
 
@@ -16,6 +18,7 @@
   title={The Lancet Commission on pollution and health},
   author={Landrigan, Philip J and Fuller, Richard and Acosta, Nereus JR and Adeyi, Olusoji and Arnold, Robert and Bald{\'e}, Abdoulaye Bibi and Bertollini, Roberto and Bose-O'Reilly, Stephan and Boufford, Jo Ivey and Breysse, Patrick N and others},
   journal={The Lancet},
+  doi={10.1016/S0140-6736(17)32345-0},
   year={2017},
   publisher={Elsevier}
 }
@@ -27,7 +30,7 @@
 	volume={3},
 	pages={1--12},
 	year={2015},
-    doi={10.3389/fenvs.2015.00009},
+  doi={10.3389/fenvs.2015.00009},
 	publisher={Frontiers}
 }
 
@@ -134,6 +137,7 @@ doi = {10.5194/gmd-2017-193}
   volume={16},
   number={2},
   pages={777--797},
+  doi={https://doi.org/10.5194/acp-16-777-2016},
   year={2016},
   publisher={Copernicus GmbH}
 }
@@ -146,7 +150,7 @@ doi = {10.5194/gmd-2017-193}
     url = {https://CRAN.R-project.org/package=covr},
   }
 @manual{cpt,
-    title = {cptcity: incorporating the cpt-city archive into R},
+    title = {cptcity: 'cpt-city' Colour Gradients},
     author = {Sergio Ibarra-Espinosa},
     year = {2017},
     note = {Colour gradients from http://soliton.vm.bytemark.co.uk/pub/cpt-city/},
