### 2018-04-25 #607 @ibarraespinosa
Hi @leeper the updated version of the [manuscript](https://github.com/atmoschem/eixport/blob/master/paper.bib) has all the respective doi.
Also, the new Zenodo DOI is [10.5281/zenodo.1228955](https://zenodo.org/record/1228955).

Many thanks!
