### 2018-04-25 095c5b9
update DOIs and URLs for JOSS
