+++ paper.bib
@@ -30,7 +30,7 @@
 	volume={3},
 	pages={1--12},
 	year={2015},
-  doi={10.3389/fenvs.2015.00009},
+        doi={10.3389/fenvs.2015.00009},
 	publisher={Frontiers}
 }
 
@@ -78,7 +78,7 @@ doi = {10.5194/gmd-2017-193}
     author = {David Pierce},
     year = {2017},
     note = {R package version 1.16},
-    url = {https://CRAN.R-project.org/package=ncdf4}
+    url = {https://cran.r-project.org/package=ncdf4}
 }
 
 @Manual{raster,
@@ -86,7 +86,7 @@ doi = {10.5194/gmd-2017-193}
     author = {Robert J. Hijmans},
     year = {2017},
     note = {R package version 2.6-7},
-    url = {https://CRAN.R-project.org/package=raster}
+    url = {https://cran.r-project.org/package=raster}
 }
 
 @article{freitas2005simple,
@@ -95,7 +95,7 @@ doi = {10.5194/gmd-2017-193}
   journal={Atmospheric Environment},
   volume={39},
   pages={6352--6361},
-  doi={https://doi.org/10.1016/j.atmosenv.2005.07.017},
+  doi={10.1016/j.atmosenv.2005.07.017},
   year={2005}
 }
 
@@ -106,7 +106,7 @@ doi = {10.5194/gmd-2017-193}
   volume={77},
   pages={748--756},
   year={2013},
-  doi={https://doi.org/10.1016/j.atmosenv.2013.05.074},
+  doi={10.1016/j.atmosenv.2013.05.074},
   publisher={Elsevier}
 }
 
@@ -118,7 +118,7 @@ doi = {10.5194/gmd-2017-193}
   number={37},
   pages={6957--6975},
   year={2005},
-  doi={https://doi.org/10.1016/j.atmosenv.2005.04.027},
+  doi={10.1016/j.atmosenv.2005.04.027},
   publisher={Elsevier}
 }
 
@@ -128,7 +128,7 @@ doi = {10.5194/gmd-2017-193}
     organization = {R Foundation for Statistical Computing},
     address = {Vienna, Austria},
     year = {2017},
-    url = {https://www.R-project.org/}
+    url = {https://www.r-project.org/}
   }
 @article{vara2016,
   title={Impact of vehicular emissions on the formation of fine particles in the Sao Paulo Metropolitan Area: a numerical study with the WRF-Chem model},
@@ -137,7 +137,7 @@ doi = {10.5194/gmd-2017-193}
   volume={16},
   number={2},
   pages={777--797},
-  doi={https://doi.org/10.5194/acp-16-777-2016},
+  doi={10.5194/acp-16-777-2016},
   year={2016},
   publisher={Copernicus GmbH}
 }
@@ -147,12 +147,12 @@ doi = {10.5194/gmd-2017-193}
     author = {Jim Hester},
     year = {2017},
     note = {R package version 3.0.1},
-    url = {https://CRAN.R-project.org/package=covr},
+    url = {https://cran.r-project.org/package=covr},
   }
 @manual{cpt,
     title = {cptcity: 'cpt-city' Colour Gradients},
     author = {Sergio Ibarra-Espinosa},
     year = {2017},
     note = {Colour gradients from http://soliton.vm.bytemark.co.uk/pub/cpt-city/},
-    url = {https://CRAN.R-project.org/package=vein},
+    url = {https://cran.r-project.org/package=vein},
   }
